ai-agent/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
openai==1.3.7
python-dotenv==1.0.0
schedule==1.2.0
pandas==2.1.3
numpy==1.25.2
scikit-learn==1.3.2
nltk==3.8.1
textblob==0.17.1
python-dateutil==2.8.2
webdriver-manager==4.0.1
fake-useragent==1.4.0
